flux-ify the sign-in logic
Before this none of our top-level (routable) components were tested and none of our Ajax code was covered. This applies the flux concept of Action -> Dispatcher -> Store -> Views for flowing data into the app. As a nice side effect it makes things very testable since each part has a reduced set of responsibilities. The testing approach relies on object API specification and mainly proves that each part adheres to the contract but this is still useful.